+++ src/components/search-results/module-classnames.ts
@@ -1,0 +1,9 @@
+import { ModuleClassnamesParams } from '@divi/module';
+
+import { SearchResultsAttrs } from './types';
+
+export const moduleClassnames = ({
+  classnamesInstance,
+}: ModuleClassnamesParams<SearchResultsAttrs>): void => {
+  classnamesInstance.add('cc-d5sr');
+};